Simplify the PM/PM2 related functions and make them more readable. Self-acked, as this was acked/committed in v2 already (r3680).
Build-tested with the AMD dbm690t target.

Acked-by: Uwe Hermann <uwe@hermann-uwe.de>



git-svn-id: svn://coreboot.org/repository/coreboot-v3@948 f3766cd6-281f-0410-b1cd-43a5c92072e9

+++ southbridge/amd/sb600/sb600.c
@@ -62,8 +62,8 @@ void set_sm_enable_bits(struct device * sm_dev, u32 reg_pos, u32 mask, u32 val)
 	}
 }
 
-void pmio_write_index(unsigned long port_base, u8 reg, u8 value);
-u8 pmio_read_index(unsigned long port_base, u8 reg);
+void pmio_write_index(u16 port_base, u8 reg, u8 value);
+u8 pmio_read_index(u16 port_base, u8 reg);
 u8 pm_ioread(u8 reg);
 void pm_iowrite(u8 reg, u8 value);
 void pm2_iowrite(u8 reg, u8 value);
